feat: User persistance across client and more

+++ client/src/util/api.js
@@ -1,30 +1,26 @@
-const axios = require("axios");
-const { API_URL } = require("./config/config");
+import axios from "axios";
 
 const api = axios.create({
-  baseURL: API_URL,
+  baseURL: "http://localhost:3001/api",
+  timeout: 1000,
+  headers: { "Content-Type": "application/json" },
   withCredentials: true,
 });
 
-//Auth
-const register = (user) => api.post("/user/register", user);
-const login = (user) => api.post("/user/auth", user);
-const logout = () => api.delete("/auth");
+//User
+export const checkAuth = () => api.get("/user/auth");
+export const logout = () => api.delete("/user/auth");
+export const login = (user) => api.post("/user/auth", user);
+export const register = (user) => api.post("/user/register", user);
 
-//Bets
-const getBets = () => api.get("/bet");
-const getBet = (id) => api.get(`/bet/${id}`);
-const createBet = (bet) => api.post("/bet", bet);
-const updateBet = (id, bet) => api.put(`/bet/${id}`, bet);
-const deleteBet = (id) => api.delete(`/bet/${id}`);
+//Bet
+export const getBets = () => api.get("/bet");
+export const createBet = (bet) => api.post("/bet", bet);
+export const getBet = (id) => api.get(`/bet/getbet/${id}`);
+export const placeBet = (id, bet) => api.post(`/bet/placebet/${id}`, bet);
 
-module.exports = {
-  register,
-  login,
-  logout,
-  getBets,
-  getBet,
-  createBet,
-  updateBet,
-  deleteBet,
-};
+//Admin
+export const updateBet = (id, bet) => api.put(`/admin/bet/${id}`, bet);
+export const deleteBet = (id) => api.delete(`/admin/bet/${id}`);
+export const updateUser = (id, user) => api.put(`/admin/user/${id}`, user);
+export const deleteUser = (id) => api.delete(`/admin/user/${id}`);
